+++ lib/DocumentActionDuplicate.js
@@ -3,7 +3,7 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.createDocumentActionDuplicate = createDocumentActionDuplicate;
+exports.DocumentActionDuplicate = DocumentActionDuplicate;
 
 var _client = _interopRequireDefault(require("part:@sanity/base/client"));
 
@@ -14,6 +14,8 @@ var _i18n = require("./i18n");
 var _router = require("@sanity/base/router");
 
 var _sanityPluginDuplicate = _interopRequireDefault(require("config:@kaliber/sanity-plugin-duplicate"));
+
+var _schema = _interopRequireDefault(require("part:@sanity/base/schema"));
 
 var _excluded = ["_id", "_createdAt", "_updatedAt"];
 
@@ -49,56 +51,69 @@ var client = _client.default.withConfig({
   apiVersion: '2022-04-06'
 });
 
-function createDocumentActionDuplicate(documentSchemes) {
-  var documentTypes = getReplacementFunctionsForAllSchemes(documentSchemes);
+function DocumentActionDuplicate(_ref) {
+  var type = _ref.type,
+      published = _ref.published,
+      draft = _ref.draft;
+  var router = (0, _router.useRouter)();
   var i18n = (0, _i18n.i18n)(_sanityPluginDuplicate.default.language);
-  return function DocumentActionDuplicate(_ref) {
-    var type = _ref.type,
-        published = _ref.published,
-        draft = _ref.draft;
-    var router = (0, _router.useRouter)();
-    return {
-      icon: _icons.CopyIcon,
-      label: 'Duplicate',
-      title: i18n['duplicate'],
-      onHandle: function () {
-        var _onHandle = _asyncToGenerator(function* () {
-          var currentDoc = draft || published;
-          if (!currentDoc) return;
+  return {
+    icon: _icons.CopyIcon,
+    label: 'Duplicate',
+    title: i18n['duplicate'],
+    onHandle: function () {
+      var _onHandle = _asyncToGenerator(function* () {
+        var currentDoc = draft || published;
+        if (!currentDoc) return;
 
-          var _id = currentDoc._id,
-              _createdAt = currentDoc._createdAt,
-              _updatedAt = currentDoc._updatedAt,
-              currentContent = _objectWithoutProperties(currentDoc, _excluded);
+        var _id = currentDoc._id,
+            _createdAt = currentDoc._createdAt,
+            _updatedAt = currentDoc._updatedAt,
+            currentContent = _objectWithoutProperties(currentDoc, _excluded);
 
-          var replacementFunctions = documentTypes === null || documentTypes === void 0 ? void 0 : documentTypes[type];
-          var replacementData = Object.fromEntries(Object.entries(replacementFunctions || []).map(_ref2 => {
-            var _ref3 = _slicedToArray(_ref2, 2),
-                fieldName = _ref3[0],
-                replacement = _ref3[1];
+        var replacementFunctions = getReplacementFunctions(_schema.default.get(type));
+        var replacementData = Object.fromEntries(Object.entries(replacementFunctions || []).map(_ref2 => {
+          var _ref3 = _slicedToArray(_ref2, 2),
+              fieldName = _ref3[0],
+              replacement = _ref3[1];
 
-            return [fieldName, typeof replacement === 'function' ? replacement(currentContent[fieldName]) : replacement];
-          }));
+          return [fieldName, typeof replacement === 'function' ? replacement(currentContent[fieldName]) : replacement];
+        }));
 
-          var doc = _objectSpread(_objectSpread(_objectSpread({}, currentContent), replacementData), {}, {
-            _id: 'drafts.'
-          });
-
-          var created = yield client.create(doc);
-          router.navigateIntent('edit', {
-            id: created._id,
-            type
-          });
+        var doc = _objectSpread(_objectSpread(_objectSpread({}, currentContent), replacementData), {}, {
+          _id: 'drafts.'
         });
 
-        function onHandle() {
-          return _onHandle.apply(this, arguments);
-        }
+        var created = yield client.create(doc);
+        router.navigateIntent('edit', {
+          id: created._id,
+          type
+        });
+      });
 
-        return onHandle;
-      }()
-    };
+      function onHandle() {
+        return _onHandle.apply(this, arguments);
+      }
+
+      return onHandle;
+    }()
   };
+}
+
+function getReplacementFunctions(schema) {
+  return Object.fromEntries(schema.fields.filter(x => {
+    var _x$type$kaliberOption;
+
+    return ![undefined, null].includes((_x$type$kaliberOption = x.type.kaliberOptions) === null || _x$type$kaliberOption === void 0 ? void 0 : _x$type$kaliberOption.duplicate);
+  }).map(field => [field.name, field.type.kaliberOptions.duplicate])); // return Object.fromEntries(schema.fields
+  //   .filter(x =>
+  //     typeof x.type.kaliberOptions?.duplicate === 'function'
+  //     || typeof x.type.kaliberOptions?.duplicate === 'string'
+  //     || typeof x.type.kaliberOptions?.duplicate === 'number'
+  //     || typeof x.type.kaliberOptions?.duplicate === 'boolean'
+  //     || typeof x.type.kaliberOptions?.duplicate === 'object'
+  //   )
+  //   .map(field => [field.name, field.type.kaliberOptions.duplicate]))
 }
 
 function getReplacementFunctionsForAllSchemes(documentSchemes) {
@@ -108,4 +123,4 @@ function getReplacementFunctionsForAllSchemes(documentSchemes) {
     return typeof ((_x$kaliberOptions = x.kaliberOptions) === null || _x$kaliberOptions === void 0 ? void 0 : _x$kaliberOptions.duplicate) === 'function' || typeof ((_x$kaliberOptions2 = x.kaliberOptions) === null || _x$kaliberOptions2 === void 0 ? void 0 : _x$kaliberOptions2.duplicate) === 'string' || typeof ((_x$kaliberOptions3 = x.kaliberOptions) === null || _x$kaliberOptions3 === void 0 ? void 0 : _x$kaliberOptions3.duplicate) === 'number' || typeof ((_x$kaliberOptions4 = x.kaliberOptions) === null || _x$kaliberOptions4 === void 0 ? void 0 : _x$kaliberOptions4.duplicate) === 'boolean' || typeof ((_x$kaliberOptions5 = x.kaliberOptions) === null || _x$kaliberOptions5 === void 0 ? void 0 : _x$kaliberOptions5.duplicate) === 'object';
   }).map(field => [field.name, field.kaliberOptions.duplicate]))]));
 }
-//# sourceMappingURL=createDocumentActionDuplicate.js.map
+//# sourceMappingURL=DocumentActionDuplicate.js.map
